+++ src/map/utils/zoneutils.cpp
@@ -474,7 +474,7 @@ namespace zoneutils
                                 PMob->m_Type        = (uint8)sql->GetIntData(22);
                                 PMob->m_Immunity    = (IMMUNITY)sql->GetIntData(23);
                                 PMob->m_EcoSystem   = (ECOSYSTEM)sql->GetIntData(24);
-                                PMob->m_ModelRadius = (uint8)sql->GetIntData(25);
+                                PMob->m_ModelRadius = (float)sql->GetIntData(25);
 
                                 PMob->speed    = (uint8)sql->GetIntData(26);
                                 PMob->speedsub = (uint8)sql->GetIntData(26);
